srv samr api conversion.  i also made SID_ENUM_TYPE parameters uint32 instead of uint8.  sorry, elrond.

+++ source/include/smb.h
@@ -384,7 +384,7 @@ typedef struct group_name_info
 	char *unix_name;
 
 	DOM_SID sid;
-	uint8 type;
+	uint32 type;
 	uint32 unix_id;
 
 } DOM_NAME_MAP;
